fix(tcpip): bound attacker-controlled reads in RTP, SMB, SIP, NDN, RADIUS, syslog (#205)
RTP: rtp_csrc_list_extraction() published cc*4 as the extracted length
but iterated it as an element count — a 176-byte heap overflow at cc=15
into the 68-byte BINARY_64DATA buffer (F-BUG-108). Iterate elements,
validate proto_offset + 12 + cc*4 <= caplen, and read each CSRC through
the unaligned-safe get_u32. mmt_rtp_search() read the sequence number
before the 12-byte minimum-length gate (F-BUG-109); the read now sits
behind the gate, and the six fixed-header bitfield extractors check the
2-byte header is captured first.

SMB: get_smb_payload() bounds the 4-byte signature compare by caplen
including after the NetBIOS skip (F-BUG-111). A new smb_payload_avail()
helper bounds every command-field read in smb_command_extraction(),
smb_transfer_payload_extraction(), smb_padding_extraction() — which
previously reached cmd_payload[48] with no check (F-BUG-112) — and
smb_session_data_analysis(). The NT_CREATE file_path_len is clamped to
the captured remainder and the allocation sized for odd lengths and
NULL-checked (F-BUG-110). smb_file_free() releases file_path->ptr and
the node itself, and smb_insert_session() takes the list head by
reference so the empty-list insert is no longer dropped (F-BUG-113).

NDN: ndn_TLV_parser() proves both header octets are captured before
reading them and accumulates the extended length in checked uint64_t
arithmetic — str_hex2int()'s pow(16, ...) int accumulation is gone and
lengths above total_length are rejected (F-BUG-065). The zero-length
root TLV no longer produces a one-past read in mmt_check_ndn_payload()
(F-BUG-076), ndn_tlv_t offsets are widened to uint32_t, every
extraction entry point clamps payload_len to caplen, and
ndn_TLV_get_string() no longer dereferences a failed str_sub().

SIP: the lowercase invite branch compared 'invite ' (7 bytes) at
offset 1 while requiring payload[7] to start 'sip:' — it could never
match (F-BUG-115); compare 'nvite ' like the uppercase branch. The
lowercase 'sip/2.0 200 OK' compare ran 14 bytes at offset 1 under a
payload_len >= 14 gate — a one-byte over-read (F-BUG-114); compare
'ip/2.0 200 OK' (13).

RADIUS: the vendor-specific field analysis read vendor_id through an
unaligned uint32_t cast — now read_be32() (F-BUG-074).

Syslog: the remaining uint8_t scan indices in the PRI parser and
classifier are widened to uint32_t, and the RFC 5424 version check no
longer reads payload[i] without an in-bounds guard off port 514
(F-BUG-055 residue).

Four new ASan/UBSan phase0 harnesses cover cc=15, the 65535-byte
declared SMB filename, truncated NDN TLV headers, and lowercase SIP
methods; the README ci-gates block is regenerated for 29 harness jobs.

+++ src/mmt_tcpip/lib/protocols/ndn.h
@@ -42,8 +42,8 @@ typedef struct ndn_tlv_struct{
 	uint16_t type;	// Type of node
 	uint8_t nb_octets; // number of octets to calculate the length of node
 	unsigned long length; // Length of node
-	uint16_t node_offset; // data offset of node in packet payload - count from type octet
-	uint16_t data_offset;
+	uint32_t node_offset; // data offset of node in packet payload - count from type octet (issue #205: uint16_t truncated offsets > 65535)
+	uint32_t data_offset;
 	struct ndn_tlv_struct *next; // sibling node - same root
 }ndn_tlv_t;
